spec: add initial model config definition (#25)
Right now we define a lot of model general information as annotations.
They should be defined as a field of model config.

We use the guidelines below to decide if a model artifact description
should be an annotation or in a structured config file:

* If a description belongs to a specific layer, it SHOULD be an annotation
* If a description describes a general property of the model, it SHOULD be
in a config

Also a `ModelFS` structure is introduced to describe the ordering of
multiple layers.

Fixes: #19
Fixes: #22

Signed-off-by: Peng Tao <bergwolf@hyper.sh>